requirements.txt
opcua
paho-mqtt
pyQt5
opcua-modeler